+++ 장우성.date/src/main/java/com/my/date/dao/UserDaoImpl.java
@@ -26,11 +26,6 @@ public class UserDaoImpl implements UserDao {
 	public User selectUserById(String id) {
 		return userMap.selectUserById(id);
 	}
-	
-	@Override
-	public String selectUserByPassword(String email) {
-		return userMap.selectUserByPassword(email);
-	}
 
 	@Override
 	public String selectEmail(String email) {
