+++ 장우성.date/src/main/java/com/my/date/dao/map/UserMap.java
@@ -8,7 +8,6 @@ public interface UserMap {
 	List<User> selectUsers();
 	User selectUser(User user);
 	User selectUserById(String id);
-	String selectUserByPassword(String email);
 	String selectEmail(String email);
 	int insertUser(User user);
 	void updateUser(User user);
