+++ 장우성.date/src/main/java/com/my/date/service/UserService.java
@@ -13,7 +13,6 @@ public interface UserService {
 	void fixUser(User user);
 	void delUser(int userId);
 	User findId(User user);
-	String getPassword(String email);
 	String getEmail(String email);
 	int fixPassword(String email, String password);
 }
